melhorias livro

+++ src/services/livro_service.go
@@ -20,31 +20,29 @@ func NewLivroService(repo *repository.LivroRepository) *LivroService {
 	}
 }
 
-func (s *LivroService) validateLivro(livro *models.Livro) error {
-	// Validações básicas usando o pacote validator
+// validateLivro: verifica se o livro possui dados válidos
+func (s *LivroService) validateLivro(livro *models.Livro) (err error) {
 	if err := s.Validator.Struct(livro); err != nil {
 		return err
 	}
 
-	// Verificação adicional: estoque não pode ser negativo
-	if livro.Estoque < 0 {
+	if livro.IsEstoqueNegativo() {
 		return errors.New("estoque não pode ser negativo")
 	}
 
-	// Verificação adicional: preço não pode ser negativo
-	if livro.Preco < 0 {
+	if livro.IsPrecoNegativo() {
 		return errors.New("preço não pode ser negativo")
 	}
 
-	// Verificação adicional: ISBN deve ser único
-	existingLivro, err := s.Repo.FindByISBN(livro.ISBN)
-	if err == nil && existingLivro.ID != 0 && existingLivro.ID != livro.ID {
+	livroBanco := s.Repo.BuscarLivro(models.Livro{ISBN: livro.ISBN})
+	if livroBanco.ID != livro.ID {
 		return errors.New("ISBN já está em uso")
 	}
 
-	return nil
+	return
 }
 
+// CreateLivro: adiciona um novo livro no banco
 func (s *LivroService) CreateLivro(livro *models.Livro) error {
 	if err := s.validateLivro(livro); err != nil {
 		return err
@@ -52,6 +50,7 @@ func (s *LivroService) CreateLivro(livro *models.Livro) error {
 	return s.Repo.Create(livro)
 }
 
+// UpdateLivro: atualiza os dados de um livro
 func (s *LivroService) UpdateLivro(livro *models.Livro) error {
 	if err := s.validateLivro(livro); err != nil {
 		return err
@@ -59,14 +58,27 @@ func (s *LivroService) UpdateLivro(livro *models.Livro) error {
 	return s.Repo.Update(livro)
 }
 
+// GetAllLivros: busca todos os livros do banco
 func (s *LivroService) GetAllLivros() ([]models.Livro, error) {
 	return s.Repo.FindAll()
 }
 
-func (s *LivroService) GetLivroByID(id uint) (*models.Livro, error) {
-	return s.Repo.FindByID(id)
+// GetLivroByID: Busca um livro pelo ID
+func (s *LivroService) GetLivroByID(id uint) *models.Livro {
+	livroParametro := models.Livro{}
+	livroParametro.ID = id
+	return s.Repo.BuscarLivro(livroParametro)
 }
 
-func (s *LivroService) DeleteLivro(id uint) error {
+// DeleteLivro: deleta um livro do banco
+func (s *LivroService) DeleteLivro(id uint) (err error) {
+	livroBanco := models.Livro{}
+	livroBanco.ID = id
+
+	livroBanco = *s.Repo.BuscarLivro(livroBanco)
+	if !livroBanco.IsID() {
+		return errors.New("livro não encontrado")
+	}
+
 	return s.Repo.Delete(id)
 }
